refactored, hopefully to improve quality..
